generating count.test
