+++ Sprint-3/revise/implement/count.js
@@ -1,0 +1,12 @@
+function countChar(str, char) {
+    let count = 0;
+    for (let i = 0; i < str.length; i++) {
+        if (str[i] === char) {
+            count++;
+        }
+    }
+    return count;
+}
+
+module.exports = countChar; // This exports the function for use in other files
+
